requirements.txt
python-dotenv==1.0.0
aiohttp==3.10.11
pyyaml==6.0.2
pydantic==2.11.4
pydantic[email]==2.11.4